api/requirements.txt

black==23.9.1 
blinker==1.6.2 
click==8.1.7 
celery==5.2.7
colorama==0.4.6 
exceptiongroup==1.1.3 
fastapi
greenlet==2.0.2 
gunicorn==21.2.0 
iniconfig==2.0.0 
isort==5.12.0 
itsdangerous==2.1.2 
jinja2==3.1.2 
markupsafe==2.1.3 
mypy-extensions==1.0.0 
mypy==1.5.1 
numpy
packaging==23.1 
pathspec==0.11.2 
pydantic
pydantic_settings
platformdirs==3.10.0 
pluggy==1.3.0 
psycopg-binary ;
psycopg2-binary
pymysql==1.1.0 
pytest==7.4.2 
python-dotenv==1.0.0 
requests
sqlmodel
sqlalchemy==2.0.20 
tomli==2.0.1 
typing-extensions==4.8.0 
werkzeug==2.3.7 
redis
